Add connectionId handling, new methods, and refactoring
Updated ChatController, FileUploadController, SessionController, ChatResponse, ChatService, LLMResponse, and SessionService to handle connectionId and new methods for deleting messages and sessions. Converted SessionItem and SessionAgent interfaces to classes. Improved chatWindow.tsx to handle new API endpoints. Removed unused import in training.tsx and updated state variables. Added new Request and Response models in both TypeScript and C#. Various formatting and minor refactoring changes for better readability and consistency.

+++ virtualteachergenaidemo.client/src/components/chat/chatWindow.tsx
@@ -6,35 +6,37 @@ import { ChatHistoryRequest, ChatMessage } from '../../models/ChatHistoryRequest
 import { HubConnectionBuilder, HubConnection } from '@microsoft/signalr';
 import { textToSpeechAsync, cancelSpeech } from '../../services/speechService';
 import { ScenarioItem } from '../../models/ScenarioItem';
-import { ISessionItem } from '../../models/SessionItem';
+import { SessionItem } from '../../models/SessionItem';
 import { Button } from '@fluentui/react-button';
-import { useUsername } from '../../auth/UserContext'; 
+import { useUsername } from '../../auth/UserContext';
+import { DeleteSessionRequest } from '../../models/Request/DeleteSessionRequest';
+import { DeleteMessageRequest } from '../../models/Request/DeleteMessageRequest';
 
 enum AuthorRole {
-    User = 0,
-    Assistant = 1
+    User,
+    Assistant
 }
 
 interface Message {
-    content: string;    
-    sessionId?: string;    
+    content: string;
+    sessionId?: string;
     id: string;
     authorRole: AuthorRole;
 }
 
 interface ChatWindowProps {
-    scenario?: ScenarioItem | undefined;
-    session?: ISessionItem;
+    scenario: ScenarioItem | null;
+    session: SessionItem | null;
 }
 
 function ChatWindow({ scenario, session }: ChatWindowProps) {
     const userName = useUsername();
     const [messages, setMessages] = useState<Message[]>([]);
-    const [connection, setConnection] = useState<HubConnection | null>(null);    
+    const [connection, setConnection] = useState<HubConnection | null>(null);
     const [sessionId, setSessionId] = useState<string>("")
     const currentMessageRef = useRef<string | null>(null);
     const [isSavingSession, setIsSavingSession] = useState<boolean>(false);
-    
+
     useEffect(() => {
         const hubUrl = process.env.HUB_URL;
         const newConnection = new HubConnectionBuilder()
@@ -49,51 +51,61 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
         if (connection) {
             connection.start()
                 .then(() => {
-                    console.log('Connected to SignalR hub');
-
-                    connection.on('ReceiveMessageUpdate', (message: any) => {
-                        if (message.state === "Start") {
-                            currentMessageRef.current = message.content;
-                            
-                            setMessages(prevMessages => [...prevMessages, {id:message.id, sessionId: message.sessionId, content: message.content, authorRole: AuthorRole.Assistant }]);
-                        } else if (message.state === "InProgress") {
-                            setMessages(prevMessages => {
-                                const updatedMessages = [...prevMessages];
-                                updatedMessages[updatedMessages.length - 1] = {id:message.id, sessionId: message.sessionId, content: message.content, authorRole: AuthorRole.Assistant }
-                                return updatedMessages;
-                            });
-                        } else if (message.state === "End") {
-                            //find message by id and update it with the id from the server
-                            setMessages(prevMessages => {
-                                const updatedMessages = [...prevMessages];
-                                updatedMessages[updatedMessages.length - 1].sessionId = message.sessionId;                                
-                                return updatedMessages;
-                            });
-                            textToSpeechAsync(message.content);
-                        }
+                    connection.on('SessionInsert', (message: any) => {
+                        setSessionId(message.sessionId);
                     });
 
-                    connection.on('UserIDUpdate', (message: any) => {
+                    connection.on('StartMessageUpdate', (message: any) => {
+                        currentMessageRef.current = message.content;
+
+                        setMessages(prevMessages => [...prevMessages, {
+                            id: message.messageId,
+                            sessionId: message.sessionId,
+                            content: message.content,
+                            authorRole: AuthorRole.Assistant
+                        }]);
+
+                        //affiche le dernier message de messages
+                        console.log("messages: ", messages[messages.length - 1]);
+                        
+                    });
+
+                    connection.on('InProgressMessageUpdate', (message: any) => {
+                        setMessages(prevMessages => {
+                                    const updatedMessages = [...prevMessages];
+                                    updatedMessages[updatedMessages.length - 1] = {id:message.id, sessionId: message.sessionId, content: message.content, authorRole: AuthorRole.Assistant }
+                                    return updatedMessages;
+                                });
+
+                    });
+
+                    connection.on('EndMessageUpdate', (message: any) => {
                         setMessages(prevMessages => {
                             const updatedMessages = [...prevMessages];
-                            const lastUserMessageIndex = updatedMessages.slice().reverse().findIndex(msg => msg.authorRole === AuthorRole.User);
-                            if (lastUserMessageIndex !== -1) {
-                                const index = updatedMessages.length - 1 - lastUserMessageIndex;                                
-                                updatedMessages[index] = { ...updatedMessages[index], sessionId: message.sessionId};
+                            updatedMessages[updatedMessages.length - 1].sessionId = message.sessionId;
+                            return updatedMessages;
+                        });
+                        textToSpeechAsync(message.content);
+
+                    });
+
+                    connection.on('MessageIdUpdate', (message: any) => {
+                        setMessages(prevMessages => {
+                            const updatedMessages = [...prevMessages];
+                            const lastUserMessage = updatedMessages.find(msg => msg.authorRole === AuthorRole.User);
+                            if (lastUserMessage !== null && lastUserMessage !== undefined) {
+                                lastUserMessage.id = message.id;                                
                             }
                             return updatedMessages;
                         });
                     });
 
-                    connection.on('SessionInsert', (message: any) => {
-                        setSessionId(message.sessionId);
-                    });
                 })
                 .catch(e => console.log('Connection failed: ', e));
         }
     }, [connection]);
 
-    useEffect(() => {        
+    useEffect(() => {
         if (session) {
             fetch(`/api/chat/messages/${session.id}`)
                 .then(response => response.json())
@@ -101,22 +113,26 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
                     const fetchedMessages = data
                         .filter((msg: Message) => msg.content !== "")
                         .map((msg: Message) => ({
-                            id: msg.sessionId,
+                            id: msg.id,
                             content: msg.content,
-                            sessionID: msg.sessionId,
+                            sessionId: msg.sessionId,
                             authorRole: msg.authorRole
-                        }));
-                    setMessages(fetchedMessages);                    
+                        })
+                        );
+                    setMessages(fetchedMessages);
+                    console.log("fetchedMessages: ", fetchedMessages);
                 })
                 .catch(error => console.error('Error fetching session messages:', error));
+
+            console.log("sessionId: ", session.id);
         }
     }, [session]);
 
 
     const handleNewMessage = async (message: string) => {
-        setMessages(prevMessages => [...prevMessages, { id: "", sessionId: sessionId, content: message, authorRole: AuthorRole.User}]);
+        setMessages(prevMessages => [...prevMessages, { id: "", sessionId: sessionId, content: message, authorRole: AuthorRole.User }]);
 
-        
+
         const currentScenario = scenario?.agents || session?.agents;
         const agent = currentScenario?.find(agent => agent.type === 'system');
         const rolePlayAgent = currentScenario?.find(agent => agent.type === 'rolePlay');
@@ -126,7 +142,7 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
         const scenarioDescription = scenario?.description || session?.scenarioDescription;
 
         //Create ISessionItem object
-        const sessionItem: ISessionItem = {
+        const sessionItem: SessionItem = {
             id: session?.id || sessionId,
             timestamp: new Date(),
             userId: userName,
@@ -152,7 +168,7 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
     const callLLMApi = async (chatHistory: ChatHistoryRequest) => {
         console.log(chatHistory);
         try {
-            const response = await fetch(`/api/chat/message?sessionId=${sessionId}`, {
+            const response = await fetch(`/api/chat/message?connectionId=${connection?.connectionId}`, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
@@ -165,19 +181,33 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
         }
     };
 
-    const handleDeleteMessage = async (index: number) => {
+    const handleDeleteMessage = async (id: string) => {
         //stop the audio
         cancelSpeech();
 
-        const messageToDelete = messages[index];
-        try {
-            await fetch(`/api/chat/messages/${messageToDelete.id}?sessionId=${messageToDelete.sessionId}`, {
-                method: 'DELETE',
-            });
-            setMessages(prevMessages => prevMessages.filter((_, i) => i !== index));
-        } catch (error) {
-            console.error('Error deleting message:', error);
+        const messageToDelete = messages.find(msg => msg.id === id);
+        if (messageToDelete !== null && messageToDelete !== undefined) {
+            const deleteRequest = new DeleteMessageRequest(messageToDelete.id, messageToDelete.sessionId!);
+
+            try {
+                const response = await fetch(`/api/chat/message/${messageToDelete.id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify(deleteRequest),
+                });
+
+                if (response.ok) {
+                    setMessages(prevMessages => prevMessages.filter(msg => msg.id !== id));
+                } else {
+                    console.error('Failed to delete message');
+                }
+            } catch (error) {
+                console.error('Error deleting message:', error);
+            }
         }
+
     };
 
     const handleSaveSession = async (sessionId: string | undefined) => {
@@ -201,6 +231,35 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
         }
     };
 
+
+    const deleteSession = async () => {
+        if (!session) return;
+
+        const deleteRequest = new DeleteSessionRequest(session.id, session.userId);
+
+        try {
+            const response = await fetch(`/api/session/delete`, {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify(deleteRequest),
+            });
+
+            if (response.ok) {
+                alert('Session deleted successfully');
+                setMessages([]);
+            } else {
+                console.error('Failed to delete session');
+                alert('Failed to delete session');
+            }
+        } catch (error) {
+            console.error('Error deleting session:', error);
+            alert('Failed to delete session');
+        }
+    };
+
+
     return (
         <div className="chat-container">
             <div className="grid-column">
@@ -217,7 +276,7 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
                             {messages.map((msg, index) => (
                                 <div key={index} className={`chat-message ${msg.authorRole === AuthorRole.User ? 'user-message' : 'assistant-message'}`}>
                                     {msg.content}
-                                    <span className="delete-message" onClick={() => handleDeleteMessage(index)}>&times;</span>
+                                    <span className="delete-message" onClick={() => handleDeleteMessage(msg.id)}>&times;</span>
                                 </div>
                             ))}
                         </div>
@@ -233,7 +292,11 @@ function ChatWindow({ scenario, session }: ChatWindowProps) {
             <div className="grid-column">
                 {/*<AudioVisualizer useMicrophone />*/}
             </div>
-            <Button appearance='primary' onClick={() =>handleSaveSession(session?.id)} disabled={isSavingSession}>Validate Session</Button>            
+            
+                <Button appearance='primary' onClick={() => handleSaveSession(session?.id)} disabled={isSavingSession}>Validate Session</Button>
+            
+                <Button appearance='secondary' onClick={() => deleteSession()}>Delete session</Button>
+            
 
         </div>
     );
